Crate fedimint_gateway_server

Re-exports

pub use config::GatewayParameters;

Modules

client

config

envs

error 🔒

events 🔒

federation_manager 🔒

iroh_server 🔒

rpc_server

types 🔒

Map gateway_lnrpc protobuf types to rust types

Structs

Gateway

Registration 🔒

Helper struct for storing the registration parameters for LNv1 for each network protocol.

Enums

GatewayState

ReceivePaymentStreamAction 🔒

The action to take after handling a payment stream.

Constants

DB_FILE 🔒

Name of the gateway’s database that is used for metadata and configuration storage.

DEFAULT_NETWORK

Default Bitcoin network for testing purposes.

DEFAULT_NUM_ROUTE_HINTS 🔒

The default number of route hints that the legacy gateway provides for invoice creation.

GW_ANNOUNCEMENT_TTL 🔒

How long a gateway announcement stays valid

LDK_NODE_DB_FOLDER 🔒

Name of the folder that the gateway uses to store its node database when running in LDK mode.

Traits

IAdminGateway

Type Aliases

AdminResult

Result