fedimint_gateway_server
Module config

Structs

GatewayOpts

Command line parameters for starting the gateway. mode, data_dir, listen, and api_addr are all required.

GatewayParameters

GatewayParameters is a helper struct that can be derived from GatewayOpts that holds the CLI or environment variables that are specified by the user.

Enums

LightningModuleMode