fedimint_gateway_server

Module envs

Constants

FM_BITCOIND_PASSWORD_ENV

FM_BITCOIND_URL_ENV

FM_BITCOIND_USERNAME_ENV

FM_DB_BACKEND_ENV

Environment variable to select database backend (rocksdb or cursed-redb)

FM_DEBUG_GATEWAY_ENV

Environment variable that instructs the gateway to run in “debug mode”, which allows errors to return to clients without redacting private information.

FM_ESPLORA_URL_ENV

FM_GATEWAY_API_ADDR_ENV

Environment variable that specifies the URL that clients can use to make requests to the gateway.

FM_GATEWAY_BCRYPT_PASSWORD_HASH_ENV

Environment variable that specifies the bcrypt password hash.

FM_GATEWAY_DATA_DIR_ENV

Environment variable that specifies the directory of the gateway’s database.

FM_GATEWAY_LIGHTNING_MODULE_MODE_ENV

Environment variable that specifies the “module mode” the gateway should run in. Options are “LNv1”, “LNv2”, or “All”. It is not recommended to run “All” in production so that clients are not able to use the same gateway to create LNv1 and LNv2 invoices.

FM_GATEWAY_LISTEN_ADDR_ENV

Environment variable that specifies the address the gateway’s HTTP server should listen on.

FM_GATEWAY_MNEMONIC_ENV

Environment variable that specifies the mnemonic that the gateway should use for ecash and the LDK Node should use for onchain funds. If not set, a mnemonic will be generated. This environment variable can be used for recovering from an existing mnemonic.

FM_GATEWAY_NETWORK_ENV

Environment variable that specifies that Bitcoin network that the gateway should use. Must match the network of the Lightning node.

FM_GATEWAY_OVERRIDE_LN_MODULE_CHECK_ENV

Environment variable that instructs the gateway to skip checking if the Bitcoin network is set to mainnet and the lightning module mode is set to All

FM_GATEWAY_SKIP_WAIT_FOR_SYNC_ENV

Environment variable that instructs the gateway to skip waiting for the bitcoin node to sync to the chain.

FM_NUMBER_OF_ROUTE_HINTS_ENV

Environment variable that instructs the gateway how many route hints to include in LNv1 invoices.