fedimint_gateway_server::gateway_new_with_builder_builder

Trait State

pub trait State: Sized {

    type BcryptPasswordHash;
    type BcryptLiquidityManagerPasswordHash;
    type GatewayState;
    type ChainSource;
    type Listen;
    type ApiAddr;
    type Network;
    type NumRouteHints;
    type DefaultRoutingFees;
    type DefaultTransactionFees;
    type IrohListen;
    type IrohDns;
    type IrohRelays;
    type MetricsListen;

}

Builder’s type state specifies if members are set or not (unset).

You can use the associated types of this trait to control the state of individual members with the IsSet and IsUnset traits. You can change the state of the members with the Set* structs available in this module.

Required Associated Types

type BcryptPasswordHash

Type state of the member bcrypt_password_hash.

It can implement either IsSet or IsUnset

type BcryptLiquidityManagerPasswordHash

Type state of the member bcrypt_liquidity_manager_password_hash.

It can implement either IsSet or IsUnset

type GatewayState

Type state of the member gateway_state.

It can implement either IsSet or IsUnset

type ChainSource

Type state of the member chain_source.

It can implement either IsSet or IsUnset

type Listen

Type state of the member listen.

It can implement either IsSet or IsUnset

type ApiAddr

Type state of the member api_addr.

It can implement either IsSet or IsUnset

type Network

Type state of the member network.

It can implement either IsSet or IsUnset

type NumRouteHints

Type state of the member num_route_hints.

It can implement either IsSet or IsUnset

type DefaultRoutingFees

Type state of the member default_routing_fees.

It can implement either IsSet or IsUnset

type DefaultTransactionFees

Type state of the member default_transaction_fees.

It can implement either IsSet or IsUnset

type IrohListen

Type state of the member iroh_listen.

It can implement either IsSet or IsUnset

type IrohDns

Type state of the member iroh_dns.

It can implement either IsSet or IsUnset

type IrohRelays

Type state of the member iroh_relays.

It can implement either IsSet or IsUnset

type MetricsListen

Type state of the member metrics_listen.

It can implement either IsSet or IsUnset

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.