fedimint_gateway_server

Module envs

Constants

FM_BITCOIND_PASSWORD_ENV

The password to use when connecting to a bitcoin node over RPC

FM_BITCOIND_URL_ENV

The URL to use when connecting to a bitcoin node over RPC. Should not include authentication parameters: (e.g http://127.0.0.1:8332)

FM_BITCOIND_USERNAME_ENV

The username to use when connecting to a bitcoin node over RPC

FM_DB_BACKEND_ENV

Environment variable to select database backend (rocksdb or cursed-redb)

FM_DEBUG_GATEWAY_ENV

Environment variable that instructs the gateway to run in “debug mode”, which allows errors to return to clients without redacting private information.

FM_DEFAULT_ROUTING_FEES_ENV

Environment variable for customizing the default routing fees

FM_DEFAULT_TRANSACTION_FEES_ENV

Environment variable for customizing the default transaction fees

FM_ESPLORA_URL_ENV

The URL to use when connecting to an Esplora server for bitcoin blockchain data

FM_GATEWAY_API_ADDR_ENV

Environment variable that specifies the URL that clients can use to make requests to the gateway.

FM_GATEWAY_BCRYPT_PASSWORD_HASH_ENV

Environment variable that specifies the bcrypt password hash.

FM_GATEWAY_DATA_DIR_ENV

Environment variable that specifies the directory of the gateway’s database.

FM_GATEWAY_LISTEN_ADDR_ENV

Environment variable that specifies the address the gateway’s HTTP server should listen on.

FM_GATEWAY_MNEMONIC_ENV

Environment variable that specifies the mnemonic that the gateway should use for ecash and the LDK Node should use for onchain funds. If not set, a mnemonic will be generated. This environment variable can be used for recovering from an existing mnemonic.

FM_GATEWAY_NETWORK_ENV

Environment variable that specifies that Bitcoin network that the gateway should use. Must match the network of the Lightning node.

FM_GATEWAY_SKIP_WAIT_FOR_SYNC_ENV

Environment variable that instructs the gateway to skip waiting for the bitcoin node to sync to the chain.

FM_NUMBER_OF_ROUTE_HINTS_ENV

Environment variable that instructs the gateway how many route hints to include in LNv1 invoices.